Fix rotation calculation for EPD modules
This code was conflating the need to adjust address offset with the rotation
calculation.  The rotation calculation should be done in a coordinate
system equal to the display height and width.  The address offset must be
done using a 'height' that represents the total number of addressable bytes
of memory configured.